Extend MultiValueMode with dedicated support of unsigned_long doc values
