+++ server/src/main/java/org/opensearch/index/fielddata/fieldcomparator/UnsignedLongValuesComparatorSource.java
@@ -24,6 +24,8 @@ import org.opensearch.index.fielddata.FieldData;
 import org.opensearch.index.fielddata.IndexFieldData;
 import org.opensearch.index.fielddata.IndexNumericFieldData;
 import org.opensearch.index.fielddata.LeafNumericFieldData;
+import org.opensearch.index.fielddata.SingletonSortedNumericUnsignedLongValues;
+import org.opensearch.index.fielddata.SortedNumericUnsignedLongValues;
 import org.opensearch.index.search.comparators.UnsignedLongComparator;
 import org.opensearch.search.DocValueFormat;
 import org.opensearch.search.MultiValueMode;
@@ -41,7 +43,6 @@ import java.math.BigInteger;
 public class UnsignedLongValuesComparatorSource extends IndexFieldData.XFieldComparatorSource {
 
     private final IndexNumericFieldData indexFieldData;
-    private final UnsignedLongMultiValueMode unsignedSortMode;
 
     public UnsignedLongValuesComparatorSource(
         IndexNumericFieldData indexFieldData,
@@ -51,7 +52,6 @@ public class UnsignedLongValuesComparatorSource extends IndexFieldData.XFieldCom
     ) {
         super(missingValue, sortMode, nested);
         this.indexFieldData = indexFieldData;
-        this.unsignedSortMode = UnsignedLongMultiValueMode.toUnsignedSortMode(sortMode);
     }
 
     @Override
@@ -59,21 +59,21 @@ public class UnsignedLongValuesComparatorSource extends IndexFieldData.XFieldCom
         return SortField.Type.LONG;
     }
 
-    private SortedNumericDocValues loadDocValues(LeafReaderContext context) {
+    private SortedNumericUnsignedLongValues loadDocValues(LeafReaderContext context) {
         final LeafNumericFieldData data = indexFieldData.load(context);
         SortedNumericDocValues values = data.getLongValues();
-        return values;
+        return new SingletonSortedNumericUnsignedLongValues(values);
     }
 
     private NumericDocValues getNumericDocValues(LeafReaderContext context, BigInteger missingValue) throws IOException {
-        final SortedNumericDocValues values = loadDocValues(context);
+        final SortedNumericUnsignedLongValues values = loadDocValues(context);
         if (nested == null) {
-            return FieldData.replaceMissing(unsignedSortMode.select(values), missingValue);
+            return FieldData.replaceMissing(sortMode.select(values), missingValue);
         }
         final BitSet rootDocs = nested.rootDocs(context);
         final DocIdSetIterator innerDocs = nested.innerDocs(context);
         final int maxChildren = nested.getNestedSort() != null ? nested.getNestedSort().getMaxChildren() : Integer.MAX_VALUE;
-        return unsignedSortMode.select(values, missingValue.longValue(), rootDocs, innerDocs, context.reader().maxDoc(), maxChildren);
+        return sortMode.select(values, missingValue.longValue(), rootDocs, innerDocs, context.reader().maxDoc(), maxChildren);
     }
 
     @Override
